config_generator/requirements.txt
netmiko
psycopg2-binary
Jinja2